+++ frontend/src/components/DevicesTable/DevicesTable.tsx
@@ -71,25 +71,39 @@ const DevicesTable: FC<DevicesTableProps> = ({refresh , setRefresh}) => {
                                     <td key={Math.random()} className='px-3 py-1 flex justify-center items-center'>
                                             <button onClick={(e)=>{
                                                     e.preventDefault();
-                                                    sendRequest({url:"api/users/device-access",method:"DELETE", params: {uuid}})
-                                                            .then(data => {
-                                                                Swal.fire({
-                                                                    position: "center",
-                                                                    icon: "success",
-                                                                    title: "Deleted Successfully",
-                                                                    showConfirmButton: false,
-                                                                    timer: 1000
-                                                                }).then(()=>setRefresh(Math.random()))
-                                                            })
-                                                            .catch(err => {
-                                                                Swal.fire({
-                                                                    position: "center",
-                                                                    icon: "error",
-                                                                    title: "can't Deleted",
-                                                                    showConfirmButton: false,
-                                                                    timer: 1000
+                                                    Swal.fire({
+                                                        title: "Are you sure?",
+                                                        text: "You won't be able to revert this!",
+                                                        icon: "warning",
+                                                        showCancelButton: true,
+                                                        confirmButtonColor: "#3085d6",
+                                                        cancelButtonColor: "#d33",
+                                                        confirmButtonText: "Yes, delete it!"
+                                                    }).then((result) => {
+                                                        if (result.isConfirmed) {
+                                                            sendRequest({url:"api/users/device-access",method:"DELETE", params: {uuid}})
+                                                                .then(data => {
+                                                                    Swal.fire({
+                                                                        position: "center",
+                                                                        icon: "success",
+                                                                        title: "Deleted Successfully",
+                                                                        showConfirmButton: false,
+                                                                        timer: 1000
+                                                                    }).then(()=>setRefresh(Math.random()))
                                                                 })
-                                                            })
+                                                                .catch(err => {
+                                                                    Swal.fire({
+                                                                        position: "center",
+                                                                        icon: "error",
+                                                                        title: "can't Deleted",
+                                                                        showConfirmButton: false,
+                                                                        timer: 1000
+                                                                    })
+                                                                })
+                                                        }
+                                                    });                
+
+                                                    
                                                                                     
                                         }} className='rounded-md bg-btns-colors-secondry min-w-[70px] md:w-2/3'>
                                                 Delete
